+++ src/main/java/domain/card/Type.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 /**
  *    카드 타입
